+++ tests/app/test_app.py
@@ -1,7 +1,7 @@
 import json
 from unittest.mock import patch
 
-from models.config import Config
+from models.config_model import Config
 
 
 def test_load_config():
@@ -105,14 +105,6 @@ def test_call_pattern_report_returns_error_when_end_date_value_is_not_a_valid_da
     response = client.get("/api/reports/call-pattern/matpal?start-date=2021-01-01&end-date=blah")
     assert response.status_code == 400
     assert response.get_data(as_text=True) == '{"error": "Invalid request, date is not valid"}'
-
-
-@patch("app.app.get_call_pattern_records_by_interviewer_and_date_range")
-def test_call_pattern_report(mock_get_call_pattern_records_by_interviewer_and_date_range, client, mock_report):
-    mock_get_call_pattern_records_by_interviewer_and_date_range.return_value = (None, 200), mock_report
-    response = client.get("/api/reports/call-pattern/matpal?start-date=2021-01-01&end-date=2021-01-01")
-    assert response.status_code == 200
-    assert response.get_data(as_text=True) == mock_report.json()
 
 
 @patch("app.app.get_call_pattern_records_by_interviewer_and_date_range")
